Create common system helpers
This moves helper methods which are used in multiple system specs in to
a single module and configures RSpec to include them in system specs. We
should only use this module for helper methods which are strictly used
in multiple system specs. I've done this to reduce the amount of
duplication in the system specs.
