+++ spec/support/system_helpers.rb
@@ -1,0 +1,22 @@
+module SystemHelpers
+  def given_the_service_is_open
+    FeatureFlag.activate(:service_open)
+  end
+
+  def given_the_service_is_closed
+    FeatureFlag.deactivate(:service_open)
+  end
+
+  def given_the_service_is_startable
+    FeatureFlag.activate(:service_start)
+  end
+
+  def when_i_am_authorized_as_a_support_user
+    page.driver.basic_authorize(
+      ENV.fetch("SUPPORT_USERNAME", "test"),
+      ENV.fetch("SUPPORT_PASSWORD", "test")
+    )
+  end
+end
+
+RSpec.configure { |config| config.include SystemHelpers, type: :system }
